Finish the header proof-of-concept
Header and navigational sidebar are set up, along with the top of the
index page for testing how everything looks. Tab navigation could still
use some work (especially in the mobile view), but I'm not experienced
enough working with that for it not to be a sub-project, so I am
leaving it to more knowledgable people.